% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hiddenFunctions.R
\name{compInverseDpss}
\alias{compInverseDpss}
\title{Composite inverse for Dpss using from schmidt hat matrix}
\usage{
compInverseDpss(
  N,
  k,
  w,
  xt,
  f,
  p,
  deltat = 1,
  polynomialOnly = FALSE,
  returnHG = FALSE
)
}
\arguments{
\item{N}{Total number of observations}

\item{k}{Number of tapers}

\item{w}{Width that gets passed into dpss()}

\item{xt}{Time series}

\item{f}{Specified frequency}

\item{p}{Highest degree polynomial}

\item{deltat}{= 1 if not specified}

\item{polynomialOnly}{If you only want the polynomial part (GH^tY == G_pC^hat_p) used in ftest}

\item{returnHG}{Will return H and G along with the Z/includes in list if poly is true}
}
\value{
Vector of compInverse for freq f, if polynomialOnly, it will return
list of cHat and polynomialPart of composit inverse which is G_p*cHat
}
\description{
Composite inverse for Dpss using from schmidt hat matrix
}
