% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hiddenFunctions.R
\name{eigenCoefdpss}
\alias{eigenCoefdpss}
\title{Non Centered eigenCoef Using DPSS}
\usage{
eigenCoefdpss(n, k, w, Xt, f, deltat = 1, dpssRet = FALSE)
}
\arguments{
\item{n}{Total number of observations}

\item{k}{Number of tapers}

\item{w}{Width that gets passed into dpss}

\item{Xt}{Time series}

\item{f}{Specific frequency you are looking at}

\item{deltat}{= 1 unless otherwise inputted}

\item{dpssRet}{= FALSE, will not return dpss use true to help not
re-computing dpss multiple times in program}
}
\value{
Vector of the eigenCoef for each k 0, ..., K-1
}
\description{
Non Centered eigenCoef Using DPSS
}
