% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hiddenFunctions.R
\name{eigenCoefdpssFFT}
\alias{eigenCoefdpssFFT}
\title{Non Centered eigenCoef Using DPSS and fft for speed}
\usage{
eigenCoefdpssFFT(
  n,
  k,
  w,
  Xt,
  deltat = 1,
  passInDPSSMat = NULL,
  pad = FALSE,
  penalty = 1,
  penaltyType = "ScaledExp"
)
}
\arguments{
\item{n}{Total number of observations}

\item{k}{Number of tapers}

\item{w}{Width that gets passed into dpss}

\item{Xt}{Time series}

\item{deltat}{= 1 by default}

\item{passInDPSSMat}{For limiting the number of calculations of DPSS this is dpss$v}

\item{pad}{if true will zero padd automatically the Xt that is inputted into the function}

\item{penalty}{1 is no penalty , 0.2  would give seq(from =  1, to =  1/(0.2*k, length.out = k)
penalty to each respective taper}
}
\value{
Matrix of EigenCoef for each frequency, as well as the frequencies under Freq in the list
}
\description{
Non Centered eigenCoef Using DPSS and fft for speed
}
