% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hiddenFunctions.R
\name{eigenCoefSine}
\alias{eigenCoefSine}
\title{Non Centered eigenCoef for Sine Tapers}
\usage{
eigenCoefSine(n, k, Xt, f, deltat = 1, sineRet = FALSE)
}
\arguments{
\item{n}{Total number of observations}

\item{k}{Number of tapers}

\item{Xt}{Time series}

\item{f}{Specific frequency you are looking at}

\item{deltat}{default is 1 unless otherwise inputted}

\item{sineRet}{= FALSE , allows for the tapers to be returned for computational efficiency}
}
\value{
Vector of the eigenCoef for each k 0, ..., K-1
}
\description{
Non Centered eigenCoef for Sine Tapers
}
