% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hiddenFunctions.R
\name{eigenCoefSineFFT}
\alias{eigenCoefSineFFT}
\title{Non Centered eigenCoef Using DPSS and fft for speed and zero padding ability}
\usage{
eigenCoefSineFFT(
  N,
  k,
  Xt,
  deltat = 1,
  passInTaper = NULL,
  returnSineMat = FALSE,
  pad = TRUE,
  penalty = 1,
  penaltyType = "ScaledExp"
)
}
\arguments{
\item{N}{Total number of observations}

\item{k}{Number of tapers}

\item{Xt}{Time series}

\item{deltat}{default is one, changes the nyquist if altered}

\item{passInTaper}{leave as null unless you are passing in a taper matrix for sine}

\item{returnSineMat}{if you want to use Sine taper matrix outside function can return if TRUE}

\item{pad}{if true, will automatically zero padd the fft}

\item{penalty}{1 is no penalty , 0.2  would give seq(from =  1, to =  1/(0.2*k, length.out = k)
penalty to each respective taper}
}
\value{
Matrix of $EigenCoef for each frequency, as well as the frequencies under $Freq in the list
if returnSineMat = TRUE, will also return $SineMat
}
\description{
Non Centered eigenCoef Using DPSS and fft for speed and zero padding ability
}
