% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hiddenFunctions.R
\name{eigenCoefSineInstFrequency}
\alias{eigenCoefSineInstFrequency}
\title{Eigen coefficients for Instantainous Frequency Sine Tapers}
\usage{
eigenCoefSineInstFrequency(
  xt,
  N,
  k,
  deltat,
  returnSineMat = FALSE,
  passInSineTapers = NULL,
  passInSineUnder = NULL,
  penalty = 1,
  penaltyOnTapersStdInv = FALSE
)
}
\arguments{
\item{xt}{time series}

\item{N}{Total number of observations}

\item{k}{Number of tapers}

\item{deltat}{default is 1, used for nyquist calculation}

\item{returnSineMat}{if you want to return SineTapers for use elsewhere}

\item{passInSineTapers}{if sine matrix was calculated somewhere else}

\item{passInSineUnder}{If sine undersampling matrix was calculated already}

\item{penalty}{1 is no penalty , 0.2  would give seq(from =  1, to =  1/(0.2*k, length.out = k)
penalty to each respective taper}

\item{penaltyOnTapersStdInv}{applys the penalty to all tapers in the calculation instead of just weighting the
eigenCoef's}
}
\value{
returns $PSI kxn matrix, $Freq which are the columns of PSI.  If
returnSineMat = TRUE, will also return tapers for future use
}
\description{
Eigen coefficients for Instantainous Frequency Sine Tapers
}
