% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hiddenFunctions.R
\name{eigenFunctionDerDpss}
\alias{eigenFunctionDerDpss}
\title{Derivative function that creates matrix of pth derivative columns and k th taper rows #'for the pth derivative with respect to f.}
\usage{
eigenFunctionDerDpss(N, k, p, w, f)
}
\arguments{
\item{N}{Total number of points in your series including zero pad}

\item{k}{number of tapers used}

\item{p}{highest derivative used}

\item{w}{bandwidth parameter, usually found with Shannons number}

\item{f}{Frequency you are looking at}
}
\value{
A matrix
}
\description{
May want to use the optimization of this in the gegenbauerPolynomialDerivation.Rmd
}
