% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hiddenFunctions.R
\name{eigenSpectrumdpssFFT}
\alias{eigenSpectrumdpssFFT}
\title{Non Centered Eigenspectrum Using DPSS and fft for speed}
\usage{
eigenSpectrumdpssFFT(
  n,
  k,
  w,
  Xt,
  deltat = 1,
  passInDPSSMat = NULL,
  pad = FALSE
)
}
\arguments{
\item{n}{Total number of observations}

\item{k}{Number of tapers}

\item{w}{Width that gets passed into dpss()}

\item{Xt}{Time series}

\item{deltat}{= 1 by default}

\item{passInDPSSMat}{For limiting the number of calculations of DPSS this is dpss()$v}
}
\value{
Matrix of EigenSpec for each frequency, as well as the frequencies under Freq in the list
}
\description{
Non Centered Eigenspectrum Using DPSS and fft for speed
}
