% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hiddenFunctions.R
\name{eigenSpectrumDPSSInstFrequency}
\alias{eigenSpectrumDPSSInstFrequency}
\title{Eigenspectra for DPSS Instantanious Frequency}
\usage{
eigenSpectrumDPSSInstFrequency(
  xt,
  N,
  k,
  w,
  deltat,
  nu = 0,
  returnDPSS = FALSE,
  passInDPSS = NULL,
  passInDPSSUnder = NULL
)
}
\arguments{
\item{xt}{time series}

\item{N}{Total number of observations}

\item{k}{Number of tapers}

\item{w}{Width that gets passed into dpss}

\item{deltat}{default is 1, used for nyquist calculation}

\item{nu}{note nu = 0 is only working right now, implimentation is there for nu != 0 but not running parallel}

\item{returnDPSS}{if you want to return dpssfor use elsewhere}

\item{passInDPSS}{if dpss has already been calculated somewhere else this is the full
object of dpss both $v and $eigen}

\item{passInDPSSUnder}{if dpss undersampling was calculated already}
}
\value{
returns $PHI kxn matrix, $Freq which are the columns of PHI.  If
returnDPSS = TRUE, will also return tapers $v and $eigen for future use
}
\description{
Eigenspectra for DPSS Instantanious Frequency
}
