% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hiddenFunctions.R
\name{eigenSpectrumSine}
\alias{eigenSpectrumSine}
\title{Non Centered Eigenspectrum for Sine Tapers}
\usage{
eigenSpectrumSine(n, k, Xt, f, deltat = 1, sineRet = FALSE)
}
\arguments{
\item{n}{Total number of observations}

\item{k}{Number of tapers}

\item{Xt}{Time series}

\item{f}{Specific frequency you are looking at}

\item{deltat}{default is 1 unless otherwise inputted}

\item{sineRet}{= FALSE , allows for the tapers to be returned for computational efficiency}
}
\value{
Vector of the eigenspectrum for each k (0, ..., K-1)
}
\description{
Non Centered Eigenspectrum for Sine Tapers
}
