% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hiddenFunctions.R
\name{eigenSpectrumSineInstFrequency}
\alias{eigenSpectrumSineInstFrequency}
\title{Eigenspectra for Instantainous Frequency Sine Tapers}
\usage{
eigenSpectrumSineInstFrequency(
  xt,
  N,
  k,
  deltat,
  nu = 0,
  returnSineMat = FALSE,
  passInSineTapers = NULL,
  passInSineUnder = NULL
)
}
\arguments{
\item{xt}{time series}

\item{N}{Total number of observations}

\item{k}{Number of tapers}

\item{deltat}{default is 1, used for nyquist calculation}

\item{nu}{= 0 default, code is in dpss version but not parallelized yet}

\item{returnSineMat}{if you want to return SineTapers for use elsewhere}

\item{passInSineTapers}{if sine matrix was calculated somewhere else}

\item{passInSineUnder}{If sine undersampling matrix was calculated already}
}
\value{
returns $PHI kxn matrix, $Freq which are the columns of PHI.  If
returnSineMat = TRUE, will also return tapers for future use
}
\description{
Eigenspectra for Instantainous Frequency Sine Tapers
}
