% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/userFunctions.R
\name{F1Test}
\alias{F1Test}
\title{F1 test statistic}
\usage{
F1Test(
  xt,
  N,
  k,
  p,
  deltat = 1,
  w = NULL,
  dpss = FALSE,
  returnInstFreqAndRegression = FALSE,
  withoutZeroDegree = FALSE
)
}
\arguments{
\item{xt}{time series}

\item{N}{Total number of observations}

\item{k}{Number of tapers}

\item{p}{Highest degree polynomial you want to test for}

\item{deltat}{Time interval between each observation}

\item{w}{Only needed if dpss = TRUE}

\item{dpss}{= FALSE unless you want to use dpss, it will do sine tapers by delfault}

\item{returnInstFreqAndRegression}{= FALSE, this speeds up the other f tests so you can pass in information}

\item{withoutZeroDegree}{TRUE if wanting modified test statistic that does not use zero degree polynomials}
}
\value{
$F1testStat and $Freq corresponding to the f1 test statistic, if reutrnInstFreqAndRegression = TRUE
it will also return $necessaryTestStuff$instFreqEigen and $necessaryTestStuff$regressionInstFreq
}
\description{
F1 test statistic
}
