% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/userFunctions.R
\name{F2Test}
\alias{F2Test}
\title{F2 test statistic}
\usage{
F2Test(
  xt,
  N,
  k,
  p,
  deltat = 1,
  w = NULL,
  dpss = FALSE,
  passInInstFreqAndRegression = NULL,
  returnInstFreqAndRegression = FALSE
)
}
\arguments{
\item{xt}{time series}

\item{N}{Total number of observations}

\item{k}{Number of tapers}

\item{p}{Highest degree polynomial you want to test for}

\item{deltat}{Time interval between each observation}

\item{w}{Only needed if dpss = TRUE}

\item{dpss}{= FALSE unless you want to use dpss, it will do sine tapers by delfault}

\item{passInInstFreqAndRegression}{leave null unless trying to simultainously do all test statistics}

\item{returnInstFreqAndRegression}{= FALSE, this speeds up the other f tests so you can pass in information}
}
\value{
$F2testStat $Freq and $necessaryTestStuff (used to pass into another ftest function)
}
\description{
F2 test statistic
}
