% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/userFunctions.R
\name{F4Prime}
\alias{F4Prime}
\title{F4 Prime}
\usage{
F4Prime(
  xt,
  k,
  p,
  N = length(xt),
  deltat = 1,
  dpss = FALSE,
  undersampleNumber = 100,
  penalty = 0.1,
  penaltyType = "ScaledExp",
  cores = 1,
  confLevel = (1 - (1/length(xt)))
)
}
\arguments{
\item{xt}{time series}

\item{k}{vector of tapers used in the f test, will conduct all at the same time.}

\item{p}{Highest degree polynomial you want to test for}

\item{N}{= length(xt)Total number of observations}

\item{deltat}{Time interval between each observation}

\item{dpss}{= FALSE unless you want to use dpss, it will do sine tapers by default}

\item{undersampleNumber}{A numeric of the number the user wants to undersample, usually 100 is a good start}

\item{penalty}{1 is no penalty , 0.1  would give seq(from =  1, to =  1/(0.1*k, length.out = k)
penalty to each respective taper}

\item{penaltyType}{What type of penalty you want to use, "ScaledExp" is the most harsh and the best right now,
"mtm" is for adaptive multitaper weighting, "Cos" is for a cosine weighting scheme, "Clip" is 1 for the number passed into penalty
k's then is 0 for the rest.  The percentage is specified by a fraction in the penalty variable}

\item{cores}{must be 1 if on windows, number of cores used for parallelization}

\item{confLevel}{default is 1-1/N, level of confidence used in the Ftest}
}
\value{
F4 list conducted at each frequency and each k.  sigFreq is a binary list of all frequencies that are considered to be significant at the given cutoff FCutOff..
}
\description{
F4 Prime
}
