% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/userFunctions.R
\name{F4Testpar}
\alias{F4Testpar}
\title{F4Test}
\usage{
F4Testpar(xt, N, k, p, deltat = 1, dpss = FALSE, undersampleNumber = 100)
}
\arguments{
\item{xt}{time series}

\item{N}{Total number of observations}

\item{p}{Highest degree polynomial you want to test for}

\item{deltat}{Time interval between each observation}

\item{dpss}{= FALSE unless you want to use dpss, it will do sine tapers by default}

\item{undersampleNumber}{A numeric of the number the user wants to undersample, usually 100 is a good start}

\item{w}{Only needed if dpss = TRUE}
}
\value{
$F4testStat, $Freq, $F14TestStat(this is the f1 not the f3 mod )
}
\description{
w is chosen by shannons number based on k
}
