% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hiddenFunctions.R
\name{FirstDerSineTaper}
\alias{FirstDerSineTaper}
\title{First Derivative of Sine Tapers}
\usage{
FirstDerSineTaper(N, k)
}
\arguments{
\item{N}{Total number of Observations}

\item{k}{Total number of tapers}
}
\value{
Matrix of evaluated derivatives for sine taper
}
\description{
First Derivative of Sine Tapers
}
