% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hiddenFunctions.R
\name{FirstDerTimeDomSlepians}
\alias{FirstDerTimeDomSlepians}
\title{First derivative of time domain slepians}
\usage{
FirstDerTimeDomSlepians(N, w, k, passInFullDPSS = NULL, returnDPSS = FALSE)
}
\arguments{
\item{N}{Total number of observations}

\item{w}{bandwidth parameter}

\item{k}{number of tapers}

\item{passInFullDPSS}{this is the object returned by dpss.  both $v and $eigen}

\item{returnDPSS}{returnDPSS for limiting calculations of dpss}
}
\value{
matrix of der for each t = 0, ..., N-1 by k = 0, ..., k-1
}
\description{
First derivative of time domain slepians
}
