% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/userFunctions.R
\name{FtestCombined}
\alias{FtestCombined}
\title{A combined F test that returns F1 modified, F3 modified and non modified (modified referring to removing the
zero polynomial test)  It is the fastest version if you want all three tests at the same time}
\usage{
FtestCombined(
  xt,
  N,
  k,
  p,
  deltat = 1,
  w = NULL,
  dpss = FALSE,
  returnInstFreqAndRegression = FALSE,
  undersample = FALSE,
  undersampleNumber = NULL
)
}
\arguments{
\item{xt}{time series}

\item{N}{Total number of observations}

\item{k}{Number of tapers}

\item{p}{Highest degree polynomial you want to test for}

\item{deltat}{Time interval between each observation}

\item{w}{Only needed if dpss = TRUE}

\item{dpss}{= FALSE unless you want to use dpss, it will do sine tapers by default}

\item{returnInstFreqAndRegression}{= FALSE, this speeds up the other f tests so you can pass in information}

\item{undersample}{True or FALSE, allows for faster run time while maintaining most accuracy, note that this will
also cause zero padding to take place.  the user DOES NOT need to manually zero pad}

\item{undersampleNumber}{A numeric of the number the user wants to undersample, usually 100 is a good start}
}
\value{
$F1Mod $F3Mod without 0Poly and $F3 with 0 poly, as well as the frequencies.  if returnInstFreq = TRUE
it will also reuturn all things needed to run the other test stats.
}
\description{
A combined F test that returns F1 modified, F3 modified and non modified (modified referring to removing the
zero polynomial test)  It is the fastest version if you want all three tests at the same time
}
