% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hiddenFunctions.R
\name{GramSchmidtMod}
\alias{GramSchmidtMod}
\title{modified gram schmidt from 2.47 and .48 and Kian Thesis}
\usage{
GramSchmidtMod(uMat, rMat)
}
\arguments{
\item{uMat}{U matrix used for the regression}

\item{rMat}{Matrix of polynomials used in regression}
}
\value{
orthonormalized U and orthogonal r
}
\description{
modified gram schmidt from 2.47 and .48 and Kian Thesis
}
