% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/userFunctions.R
\name{grumbleModulationGeneration}
\alias{grumbleModulationGeneration}
\title{Modulation Generation for grumble distribution}
\usage{
grumbleModulationGeneration(
  N,
  P,
  BLinear,
  linConstCoef,
  linCoef,
  AmpLinear,
  BQuadratic = 0,
  quadConstCoef = 0,
  quadLinCoef = 0,
  quadCoef = 0,
  AmpQuad = 0,
  BCubic = 0,
  cubeConstCoef = 0,
  cubeLinCoef = 0,
  cubeQuadCoef = 0,
  cubeCoef = 0,
  AmpCube = 0,
  BQuartic = 0,
  quartConstCoef = 0,
  quartLinCoef = 0,
  quartQuadCoef = 0,
  quartCubeCoef = 0,
  quartCoef = 0,
  AmpQuart = 0,
  fLin = 0.1,
  fQuad = 0.3,
  fCube = 0.31,
  fQuart = 0.4,
  checkBandWidth = FALSE,
  ar = c(0.5, 0.3, -0.1),
  ma = c(0.6),
  noiseScale = 1 * 6/pi^2,
  plotXt = FALSE
)
}
\arguments{
\item{N}{Length of series you want to create}

\item{P}{Highest degree polynomial will use all degrees less than P other than 0}

\item{BLinear}{desired bandwidth for linear modulation}

\item{linConstCoef}{linear constant term}

\item{linCoef}{linear term coefficient}

\item{AmpLinear}{amplitude of linear modulation}

\item{BQuadratic}{desired bandwidth for quadratic modulation}

\item{quadConstCoef}{quadratic constant term}

\item{quadLinCoef}{quadratic linear term coefficient}

\item{quadCoef}{quadratic term coefficient}

\item{AmpQuad}{amplitude of quadratic modulation}

\item{BCubic}{desired bandwidth for cubic modulation}

\item{cubeConstCoef}{cubic constant term}

\item{cubeLinCoef}{cubic linear term coefficient}

\item{cubeQuadCoef}{cubic quadratic term coefficient}

\item{cubeCoef}{cubic term coefficient}

\item{AmpCube}{amplitide of cubic modulation}

\item{BQuartic}{desired bandwidth for quartic modulation}

\item{quartConstCoef}{quartic constant term}

\item{quartLinCoef}{quartic linear term coefficient}

\item{quartQuadCoef}{quartic quadratic term coefficient}

\item{quartCubeCoef}{quartic cubic term coefficient}

\item{quartCoef}{quartic term coefficient}

\item{AmpQuart}{amplitude of quartic modulation}

\item{fLin}{desired linear modulation carrier frequency}

\item{fQuad}{desired quadratic modulation carrier frequency}

\item{fCube}{desired cubic modulation carrier frequency}

\item{fQuart}{desired quartic modulation carrier frequency}

\item{checkBandWidth}{def = FALSE  if TRUE will print out the bandwidths to check}

\item{ar}{=  0.5, 0.3, -0.1 AR coef for noise generation}

\item{ma}{= 0.6 MA coef for noise generation}

\item{noiseScale}{= 1*6/pi^2 ratio of noise to pure signal}

\item{plotXt}{def = FALSE, will plot xt with noise if needed}
}
\value{
$xt for noisy data and $xtNoNoise for signal without noise and $noise for just the pure noise that was used
}
\description{
Modulation Generation for grumble distribution
}
