% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hiddenFunctions.R
\name{HatMatGMatDpss}
\alias{HatMatGMatDpss}
\title{GramShmidt Orthonormalized Hat and Polynomial Matrices for Dpss}
\usage{
HatMatGMatDpss(N, k, p, w)
}
\arguments{
\item{N}{Total Number of Observations}

\item{k}{Number of tapers}

\item{p}{Highest degree polynomial}

\item{w}{Band Width parameter passed into dpss}
}
\value{
List of H hat matrix and G polynomial matrix
}
\description{
GramShmidt Orthonormalized Hat and Polynomial Matrices for Dpss
}
