% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hiddenFunctions.R
\name{instFreqDPSS}
\alias{instFreqDPSS}
\title{Instantaneous Frequency for DPSS}
\usage{
instFreqDPSS(
  xt,
  N,
  k,
  w,
  deltat,
  passInDPSS = NULL,
  returnDPSS = FALSE,
  passInDPSSUnder = NULL,
  penalty = 1,
  penaltyType = "ScaledExp",
  pad = TRUE
)
}
\arguments{
\item{xt}{time series}

\item{N}{Total number of observations}

\item{k}{Number of tapers}

\item{w}{Width that gets passed into dpss}

\item{deltat}{default is 1, used for nyquist calculation}

\item{passInDPSS}{if dpss has already been calculated somewhere else this is the full
object of dpss both $v and $eigen}

\item{returnDPSS}{if you want to return dpss for use elsewhere}

\item{passInDPSSUnder}{if dpss undersampling was calculated already}

\item{penalty}{1 is no penalty , 0.2  would give seq(from =  1, to =  1/(0.2*k, length.out = k)
penalty to each respective taper}
}
\value{
list of $InstFreq and $Freq which is the column of instFreq
if returnDPSS = TRUE then list of $InstFreq, $Freq, and $DPSS
}
\description{
Instantaneous Frequency for DPSS
}
