% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hiddenFunctions.R
\name{instFreqSine}
\alias{instFreqSine}
\title{Instantaneous Frequency for Sine}
\usage{
instFreqSine(
  xt,
  N,
  k,
  deltat,
  passInSineMat = NULL,
  returnSineMat = FALSE,
  passInSineUnder = NULL,
  penalty = 1,
  penaltyType = "ScaledExp",
  penaltyOnTapersStdInv = FALSE,
  reduced = FALSE,
  pad = TRUE
)
}
\arguments{
\item{xt}{time series}

\item{N}{Total number of observations}

\item{k}{Number of tapers}

\item{deltat}{default is 1, used for nyquist calculation}

\item{passInSineMat}{if sine matrix was calculated somewhere else}

\item{returnSineMat}{if you want to return SineTapers for use elsewhere}

\item{passInSineUnder}{If sine undersampling matrix was calculated already}

\item{penalty}{1 is no penalty , 0.2  would give seq(from =  1, to =  1/(0.2*k, length.out = k)
penalty to each respective taper}

\item{penaltyType}{= "ScaledExp" by default, allows for others to be created in the future}

\item{penaltyOnTapersStdInv}{applys the penalty to all tapers in the calculation instead of just weighting the
eigenCoef's}

\item{reduced}{used when prime test is conducted}

\item{pad}{= TRUE, but allows for user setable if they want a different amount of padding}
}
\value{
list of $InstFreq and $Freq which is the column of instFreq
if returnSineMat = TRUE then list of $InstFreq, $Freq, and $SineMat
}
\description{
Instantaneous Frequency for Sine
}
