% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hiddenFunctions.R
\name{regressionDPSSInstFreq}
\alias{regressionDPSSInstFreq}
\title{regression for DPSS Tapers using Gram schmidt hat matrix specifically for Inst Freq}
\usage{
regressionDPSSInstFreq(
  N,
  k,
  w,
  instFreqEigen,
  p,
  passInDPSS = NULL,
  returnDPSS = FALSE,
  returnRp = FALSE,
  withoutzeroPoly = FALSE,
  returnGCHatp = FALSE
)
}
\arguments{
\item{N}{Total number of observations}

\item{k}{Number of tapers}

\item{w}{bandwidth for the multitaper}

\item{instFreqEigen}{inst frequency eigen spectrum}

\item{p}{Highest degree polynomial}

\item{passInDPSS}{if you have already computed dpss() with $v and $eigen}

\item{returnDPSS}{if you need the tapers for other functions}

\item{returnRp}{will return the residuals if needed}

\item{withoutzeroPoly}{used if using modified f tests without 0th order polynomial}

\item{returnGCHatp}{will return GCHatP if needed}
}
\value{
returnlist of $cHat and polynomialPart of composit inverse $GCHatp, $H and $G, \
if returnDPSS = TRUE then will return tapers as well $DPSS
}
\description{
regression for DPSS Tapers using Gram schmidt hat matrix specifically for Inst Freq
}
