% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hiddenFunctions.R
\name{regressionSineInstFreq}
\alias{regressionSineInstFreq}
\title{regression for Sine Tapers using Gram schmidt hat matrix specifically for Inst Freq}
\usage{
regressionSineInstFreq(
  N,
  k,
  instFreqEigen,
  p,
  passInSineMat = NULL,
  returnSineTapers = FALSE,
  returnRp = FALSE,
  withoutzeroPoly = FALSE,
  returnGCHatp = FALSE
)
}
\arguments{
\item{N}{Total number of observations}

\item{k}{Number of tapers}

\item{instFreqEigen}{inst frequency eigen spectrum}

\item{p}{Highest degree polynomial}

\item{passInSineMat}{if you have already computed the matrix of tapers}

\item{returnSineTapers}{if you need the tapers for other functions}

\item{returnRp}{if you want residuals to be returned and calculated as well}

\item{withoutzeroPoly}{if you dont want zeroth order polynomials to be in test statistics}
}
\value{
returnlist of $cHat and polynomialPart of composit inverse $GCHatp, $H and $G, \
if returnSineTapers = TRUE then will return tapers as well $sineTapers
}
\description{
regression for Sine Tapers using Gram schmidt hat matrix specifically for Inst Freq
}
