% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hiddenFunctions.R
\name{Rnp}
\alias{Rnp}
\title{Matrix whos pth column is the unscaled monomial of degree p on the interval -1 to 1}
\usage{
Rnp(n, p, N)
}
\arguments{
\item{n}{the nth point in the series 0 to N-1}

\item{p}{the degree of the polynomial 0 to P}

\item{N}{The total number of observations}
}
\value{
signle n,pth element from the matrix
}
\description{
Matrix whos pth column is the unscaled monomial of degree p on the interval -1 to 1
}
