% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hiddenFunctions.R
\name{RnpMat}
\alias{RnpMat}
\title{Polynomial matrix generation}
\usage{
RnpMat(N, P)
}
\arguments{
\item{N}{total number of observations}

\item{P}{highest degree polynomial}
}
\value{
matrix of all polynomial numbers
}
\description{
Polynomial matrix generation
}
