% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hiddenFunctions.R
\name{sineTaperMatrix}
\alias{sineTaperMatrix}
\title{Sinusoidal Taper Matrix}
\usage{
sineTaperMatrix(N, k)
}
\arguments{
\item{N}{Length of Time series}

\item{k}{Number of tapers starting at 0}
}
\value{
Matrix of tapers for the given N and k values, N columns starting at n = 0
}
\description{
This is the same as the one Ben and Yatharth added to multitaper but is returned in the same orientation as the one used in Percival and Walden
}
