% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hiddenFunctions.R
\name{singleIterationForParallel}
\alias{singleIterationForParallel}
\title{Single iteration of F3Mod for Parallel applications}
\usage{
singleIterationForParallel(
  xt,
  k,
  p,
  deltat = 1,
  w = NULL,
  dpss = FALSE,
  undersampleNumber = 100,
  confLevel = (1 - (1/length(xt))),
  returnFTestVars = FALSE,
  penalty = 1
)
}
\arguments{
\item{xt}{vector of time series observations}

\item{k}{single k that will be used for the f test}

\item{p}{largest degree of the polynomial modulation suspected}

\item{deltat}{= 1 by default}

\item{w}{multitaper bandwidth parameter}

\item{dpss}{= FALSE uses sine instead}

\item{undersampleNumber}{= 100 by default  this is the number used in undersampling the tapers}

\item{confLevel}{level of confidence used in F3Mod, default is 1-1/N}

\item{returnFTestVars}{= FALSE.  Used for when more information about inner Ftests is needed. NOTE: this will be very memory intensive}

\item{penalty}{1 is no penalty , 0.2  would give seq(from =  1, to =  1/(0.2*k, length.out = k)
penalty to each respective taper}
}
\value{
$F3Mod, $Freq, $significantFreq, $k
}
\description{
can be used inside mclapply.  note that there is no checks in here as they are assumed
to be ran in the parent user function
}
