% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hiddenFunctions.R
\name{singleIterationForParallel4}
\alias{singleIterationForParallel4}
\title{Single iteration of F4Mod for Parallel applications}
\usage{
singleIterationForParallel4(
  xt,
  N,
  k,
  p,
  deltat = 1,
  w = NULL,
  dpss = FALSE,
  undersampleNumber = 100
)
}
\arguments{
\item{xt}{vector of time series observations}

\item{N}{length of Xt}

\item{k}{single k that will be used for the f test}

\item{p}{largest degree of the polynomial modulation suspected}

\item{deltat}{= 1 by default}

\item{w}{multitaper bandwidth parameter}

\item{dpss}{= FALSE uses sine instead}

\item{undersampleNumber}{= 100 by default  this is the number used in undersampling the tapers}
}
\value{
$cHat chat used in caluclation of f , $PHI capital PHI , $Freq vector of frequencies , $k number of tapers used
}
\description{
can be used inside mclapply.  note that there is no checks in here as they are assumed
to be ran in the parent user function
}
