% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hiddenFunctions.R
\name{singleIterationForParallelAllTypeSwitcher}
\alias{singleIterationForParallelAllTypeSwitcher}
\title{Switch between F4 and FPrime tests, used in agg test.}
\usage{
singleIterationForParallelAllTypeSwitcher(
  xt,
  k,
  p = p,
  deltat = 1,
  FPrime = TRUE,
  undersampleNumber = 100,
  confLevel = (1 - (1/length(xt))),
  dpss = FALSE,
  penalty = 1,
  penaltyType = "ScaledExp"
)
}
\arguments{
\item{xt}{time series input}

\item{k}{number of tapers that one wants to use}

\item{p}{what degree polynomial are we testing for}

\item{deltat}{= 1 by default but can be changed}

\item{FPrime}{= TRUE, if so it uses the reduced test}

\item{undersampleNumber}{= 100, this is usually a good starting point}

\item{confLevel}{= 1-1/N by default}

\item{dpss}{= FALSE, if TRUE and using FPrime, it will still use sine tapers for the even K's as that is the only prime test set up}

\item{penalty}{= 0.15, this is usually a good starting point, may need to be increased for prime test}

\item{penaltyType}{= "ScaledExp", this is the best one at this point}
}
\value{
see return for singleIterationForParallel
}
\description{
Switch between F4 and FPrime tests, used in agg test.
}
