% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hiddenFunctions.R
\name{singleIterationForParallelFPrime}
\alias{singleIterationForParallelFPrime}
\title{Single iteration for the FPrime test, set up to use mclapply to run more than one k at a time if wanting to}
\usage{
singleIterationForParallelFPrime(
  xt,
  k,
  p = p,
  deltat = 1,
  undersampleNumber = 100,
  confLevel = (1 - (1/length(xt))),
  dpss = FALSE,
  penalty = 1,
  penaltyType = "ScaledExp"
)
}
\arguments{
\item{xt}{time series input}

\item{k}{number of tapers that one wants to use}

\item{p}{what degree polynomial are we testing for}

\item{deltat}{= 1 by default but can be changed}

\item{undersampleNumber}{= 100, this is usually a good starting point}

\item{confLevel}{= 1-1/N by default}

\item{dpss}{= FALSE, if TRUE and using FPrime, it will still use sine tapers for the even K's as that is the only prime test set up}

\item{penalty}{= 0.15, this is usually a good starting point, may need to be increased for prime test}

\item{penaltyType}{= "ScaledExp", this is the best one at this point}
}
\value{
returns the Ftest, Freq it tested at, significantFrequencies, and the cut off FCutOff for the specific test
}
\description{
Single iteration for the FPrime test, set up to use mclapply to run more than one k at a time if wanting to
}
