% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hiddenFunctions.R
\name{sinTaperSingle}
\alias{sinTaperSingle}
\title{Single Sinusoidal Taper}
\usage{
sinTaperSingle(N, k, t)
}
\arguments{
\item{N}{Length of time series}

\item{k}{Number of tapers}

\item{t}{Specific time point that you are interested in t starts at zero}
}
\value{
Only the one sin taper that you are interested in at the single point
}
\description{
This returns only the one sin taper that you are interested in at the single point
}
