% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hiddenFunctions.R
\name{standardInverseDPSSSingleFreq}
\alias{standardInverseDPSSSingleFreq}
\title{Standard Inversion for Polynomial Projection Filter DPSS at freq f}
\usage{
standardInverseDPSSSingleFreq(
  xt,
  N,
  w,
  k,
  f,
  deltat = 1,
  FFT = FALSE,
  retDPSS = FALSE
)
}
\arguments{
\item{xt}{time series}

\item{N}{Total number of observations}

\item{w}{Width that gets passed into dpss()}

\item{k}{Number of tapers}

\item{f}{specified frequency you want to look at}

\item{deltat}{= 1 if not specified, but is only used if not using FFT}

\item{FFT}{if you want to use fft() to speed up or not have delta t = 1 with zero padding}

\item{retDPSS}{if TRUE, returns calculated DPSS for pass in to other functions}
}
\value{
Standard Inverse vector for frequency f
}
\description{
Standard Inversion for Polynomial Projection Filter DPSS at freq f
}
