% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hiddenFunctions.R
\name{standardInverseSineDer}
\alias{standardInverseSineDer}
\title{Standard Inversion first Derivative for Polynomial Projection Filter Sine}
\usage{
standardInverseSineDer(
  xt,
  N,
  k,
  deltat = 1,
  passInSineMat = NULL,
  returnSineMat = FALSE,
  passInSineUnder = NULL
)
}
\arguments{
\item{xt}{time series}

\item{N}{Total number of observations}

\item{k}{Number of tapers}

\item{deltat}{default is 1 unless using different nyquist}

\item{FirstDir}{= NULL, allows for the derivative for the sine taper to be passed in for faster calculation}
}
\value{
list of $StdInverse and $Freq which are the columns of StdInverse,
if returnSineMat = TRUE then will also return $SineMat
}
\description{
NOTE: this could be sped up by only calculating the fft once!! for all f
}
