% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hiddenFunctions.R
\name{standardInverseSineSingleFreq}
\alias{standardInverseSineSingleFreq}
\title{Standard Inversion for Polynomial Projection Filter Sine for a freq f}
\usage{
standardInverseSineSingleFreq(xt, N, k, f, deltat = 1, FFT = FALSE)
}
\arguments{
\item{xt}{time series}

\item{N}{Total number of observations}

\item{k}{Number of tapers}

\item{f}{specified frequency you want to look at}

\item{deltat}{= 1 if not specified, but is only used if not using FFT}

\item{FFT}{if you want to use fft to speed up or not have delta t = 1 with zero padding}
}
\value{
Standard Inverse vector for frequency f
}
\description{
Standard Inversion for Polynomial Projection Filter Sine for a freq f
}
