% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hiddenFunctions.R
\name{UDpss}
\alias{UDpss}
\title{Matrix of Ukp's defined by 2.36 in Kians thesis for the DPSS}
\usage{
UDpss(N, k, p, w, round = 14, passInDPSS = NULL)
}
\arguments{
\item{N}{Total number of points in your series including zero pad}

\item{k}{number of tapers used 0 to k-1}

\item{p}{highest derivative used/highest degree polynomial starts at 0}

\item{w}{bandwidth parameter, usually found with Shannons number}

\item{round}{if 16 you will obtain not zero in k != p mod 2}
}
\value{
matrix
}
\description{
Matrix of Ukp's defined by 2.36 in Kians thesis for the DPSS
}
