% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hiddenFunctions.R
\name{USine}
\alias{USine}
\title{U_kp used in the regression for sine tapers}
\usage{
USine(N, k, p, round = 15, passInSineTapers = NULL)
}
\arguments{
\item{N}{Total number of observations}

\item{k}{the number of tapers 0 to k-1}

\item{p}{Highest degree der used/ highest degree polynomial 0 to p}

\item{round}{default 15 decimals to obtain exactly 0 in the non mod 2 points,
I believe it has to do with the rounding on pi for why its not exactly zero.}

\item{passInSineTapers}{used for speeding up the calculation.}
}
\value{
matrix
}
\description{
U_kp used in the regression for sine tapers
}
