% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/userFunctions.R
\name{F3Testpar}
\alias{F3Testpar}
\title{F3TestParallel}
\usage{
F3Testpar(
  xt,
  N = length(xt),
  k,
  p,
  deltat = 1,
  dpss = FALSE,
  undersampleNumber = 100,
  cores = 1,
  confLevel = (1 - (1/length(xt)))
)
}
\arguments{
\item{xt}{time series}

\item{N}{Total number of observations}

\item{k}{vector of tapers used in the f test}

\item{p}{Highest degree polynomial you want to test for}

\item{deltat}{Time interval between each observation}

\item{dpss}{= FALSE unless you want to use dpss, it will do sine tapers by default}

\item{undersampleNumber}{A numeric of the number the user wants to undersample, usually 100 is a good start}

\item{cores}{must be 1 if on windows, number of cores used for parallelization}
}
\value{
$F3testStat, $Freq, $significantFrequencies
}
\description{
w is chosen by shannons number based on k
}
