% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/userFunctions.R
\name{F4Test}
\alias{F4Test}
\title{F4Test}
\usage{
F4Test(
  xt,
  N,
  k,
  p,
  deltat = 1,
  w = NULL,
  dpss = FALSE,
  undersampleNumber = 100
)
}
\arguments{
\item{xt}{time series}

\item{N}{Total number of observations}

\item{k}{Number of tapers}

\item{p}{Highest degree polynomial you want to test for}

\item{deltat}{Time interval between each observation}

\item{w}{Only needed if dpss = TRUE}

\item{dpss}{= FALSE unless you want to use dpss, it will do sine tapers by default}

\item{undersampleNumber}{A numeric of the number the user wants to undersample, usually 100 is a good start}

\item{returnInstFreqAndRegression}{= FALSE, this speeds up the other f tests so you can pass in information}

\item{withoutZeroDegree}{TRUE if wanting modified test statistic that does not use zero degree polynomials (no undersampling if FALSE)}

\item{undersample}{True or FALSE, allows for faster run time while maintaining most accuracy, note that this will
also cause zero padding to take place.  the user DOES NOT need to manually zero pad}
}
\value{
$F4testStat, $Freq, $necessaryTestStuff (used to pass into another ftest function)
}
\description{
F4Test
}
