% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/userFunctions.R
\name{gumbelModulationGeneration}
\alias{gumbelModulationGeneration}
\title{Modulation Generation for gumble distribution}
\usage{
gumbelModulationGeneration(
  N,
  P,
  wLinear,
  linCoefs,
  AmpLinear,
  wQuadratic = 0,
  quadCoefs = c(0, 0, 0),
  AmpQuad = 0,
  wCubic = 0,
  cubeCoefs = c(0, 0, 0, 0),
  AmpCube = 0,
  wQuartic = 0,
  quartCoefs = c(0, 0, 0, 0, 0),
  AmpQuart = 0,
  fLin = 0.1,
  fQuad = 0.3,
  fCube = 0.31,
  fQuart = 0.4,
  ar = c(0.5, 0.3, -0.1),
  ma = c(0.6),
  noiseScale = 1 * 6/pi^2,
  checkBandWidth = FALSE,
  plotXt = FALSE,
  seed = NULL
)
}
\arguments{
\item{N}{Length of series you want to create}

\item{P}{Highest degree polynomial will use all degrees less than P other than 0}

\item{wLinear}{desired bandwidth for linear modulation}

\item{linCoefs}{vector of c1 and c2 in \eqn{BLinear\cdot(c1 + c2 \cdot t)}}

\item{AmpLinear}{amplitude of linear modulation}

\item{wQuadratic}{desired bandwidth for quadratic modulation}

\item{quadCoefs}{vector of c1, c2, c3 in \eqn{BQuadratic\cdot(c1 + c2 \cdot t + c3 \cdot t^2)}}

\item{AmpQuad}{amplitude of quadratic modulation}

\item{wCubic}{desired bandwidth for cubic modulation}

\item{cubeCoefs}{vector of c1, c2, c3, c4 in \eqn{BCubic\cdot(c1 + c2 \cdot t + c3 \cdot t^2 + c4 \cdot t^3)}}

\item{AmpCube}{amplitide of cubic modulation}

\item{wQuartic}{desired bandwidth for quartic modulation}

\item{quartCoefs}{vector of c1, c2, c3, c4, c5 in \eqn{BQuart\cdot(c1 + c2 \cdot t + c3 \cdot t^2 + c4 \cdot t^3 + c5 \cdot^4)}}

\item{AmpQuart}{amplitude of quartic modulation}

\item{fLin}{desired linear modulation carrier frequency}

\item{fQuad}{desired quadratic modulation carrier frequency}

\item{fCube}{desired cubic modulation carrier frequency}

\item{fQuart}{desired quartic modulation carrier frequency}

\item{ar}{=  0.5, 0.3, -0.1 AR coef for noise generation}

\item{ma}{= 0.6 MA coef for noise generation}

\item{noiseScale}{= 1*6/pi^2 ratio of noise to pure signal}

\item{checkBandWidth}{def = FALSE  if TRUE will print out the bandwidths to check}

\item{plotXt}{def = FALSE, will plot xt with noise if needed}

\item{seed}{seed used for generation of the gumbel noise}
}
\value{
$xt for noisy data and $xtNoNoise for signal without noise and $noise for just the pure noise that was used
        $correctionCoef tells you the coeficients that give the correct wp's (Blinear, ...., BQuartic)
}
\description{
Modulation Generation for gumble distribution
}
