% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hiddenFunctions.R
\name{HatMatGMatSine}
\alias{HatMatGMatSine}
\title{GramShmidt Orthonormalized Hat and Polynomial Matrices for Sine}
\usage{
HatMatGMatSine(N, k, p, passInSineTapers = NULL)
}
\arguments{
\item{N}{Total Number of Observations}

\item{k}{Number of tapers}

\item{p}{Highest degree polynomial}

\item{passInSineTapers}{used for speeding up the calculation.}
}
\value{
List of H hat matrix and G polynomial matrix
}
\description{
GramShmidt Orthonormalized Hat and Polynomial Matrices for Sine
}
