% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hiddenFunctions.R
\name{standardInverseDPSS}
\alias{standardInverseDPSS}
\title{Standard Inverse with fft for all frequencies using DPSS}
\usage{
standardInverseDPSS(
  xt,
  N,
  k,
  w,
  deltat = 1,
  retDPSS = FALSE,
  passInDPSS = NULL,
  passInDPSSReduced = NULL,
  penalty = 1,
  penaltyType = "ScaledExp"
)
}
\arguments{
\item{xt}{time series}

\item{N}{Total number of observations}

\item{k}{Number of tapers}

\item{w}{Width that gets passed into dpss}

\item{deltat}{default is 1, used for nyquist calculation}

\item{retDPSS}{If true, will return the dpss for pass into other functions}

\item{passInDPSS}{if dpss has already been calculated somewhere else this is the full
object of dpss both $v and $eigen}

\item{passInDPSSReduced}{if dpss undersampling was calculated already}

\item{penalty}{1 is no penalty , 0.2  would give seq(from =  1, to =  1/(0.2*k, length.out = k)
penalty to each respective taper}
}
\value{
returns standard inverse along with frequencies that are the columns
}
\description{
Standard Inverse with fft for all frequencies using DPSS
}
