% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hiddenFunctions.R
\name{standardInverseDPSSFirstDir}
\alias{standardInverseDPSSFirstDir}
\title{Standard Inversion for Polynomial Projection Filter First Derivative DPSS}
\usage{
standardInverseDPSSFirstDir(
  xt,
  N,
  w,
  k,
  deltat = 1,
  passInDPSS = NULL,
  returnDPSS = FALSE,
  passInDPSSReduced = NULL
)
}
\arguments{
\item{xt}{time series}

\item{N}{Total number of observations}

\item{w}{Width that gets passed into dpss}

\item{k}{Number of tapers}

\item{deltat}{default is 1 used for nyquist}

\item{passInDPSS}{used for speed if you have already calculated the dpss object somewhere else
this includes both $v and $eigen}

\item{returnDPSS}{returns dpss if needed outside function}

\item{passInDPSSReduced}{if dpss undersampling was calculated already}
}
\value{
Standard Inverse Derivative Z' vector $StdInverse as well as $Freq which are the columns of stdInverse
if returnDPSS = TRUE will also return $DPSS which is the dpss object
}
\description{
Standard Inversion for Polynomial Projection Filter First Derivative DPSS
}
