% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hiddenFunctions.R
\name{standardInverseSine}
\alias{standardInverseSine}
\title{Standard Inverse for Sine tapers}
\usage{
standardInverseSine(
  xt,
  N,
  k,
  deltat = 1,
  passInSineMat = NULL,
  retSineTapers = FALSE,
  passInSineUnder = NULL
)
}
\arguments{
\item{xt}{time series}

\item{N}{Total number of observations}

\item{k}{Number of tapers}

\item{deltat}{default is 1 unless using different nyquist}

\item{passInSineMat}{If sine matrix was calculated already}

\item{retSineTapers}{Will return calculated sine taper matrix for future use}
}
\value{
returns matrix of standard inverse along with frequencies corresponding to the columns
}
\description{
Standard Inverse for Sine tapers
}
