% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hiddenFunctions.R
\name{standardInverseSine}
\alias{standardInverseSine}
\title{Standard Inverse for Sine tapers}
\usage{
standardInverseSine(
  xt,
  N,
  k,
  deltat = 1,
  passInSineMat = NULL,
  retSineTapers = FALSE,
  passInSineUnder = NULL,
  penalty = 1,
  penaltyType = "ScaledExp",
  penaltyOnTapersStdInv = FALSE,
  reduced = FALSE,
  pad = TRUE
)
}
\arguments{
\item{xt}{time series}

\item{N}{Total number of observations}

\item{k}{Number of tapers}

\item{deltat}{default is 1 unless using different nyquist}

\item{passInSineMat}{If sine matrix was calculated already}

\item{retSineTapers}{Will return calculated sine taper matrix for future use}

\item{passInSineUnder}{If sine undersampling matrix was calculated already}

\item{penalty}{1 is no penalty , 0.2  would give seq(from =  1, to =  1/(0.2*k, length.out = k)
penalty to each respective taper}

\item{penaltyType}{= "ScaledExp" by default, allows for others to be created in the future}

\item{penaltyOnTapersStdInv}{applys the penalty to all tapers in the calculation instead of just weighting the
eigenCoef's}

\item{reduced}{used when prime test is conducted}

\item{pad}{= TRUE, but allows for user setable if they want a different amount of padding}
}
\value{
returns matrix of standard inverse along with frequencies corresponding to the columns
}
\description{
Standard Inverse for Sine tapers
}
