% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hiddenFunctions.R
\name{standardInverseSineDer}
\alias{standardInverseSineDer}
\title{Standard Inversion first Derivative for Polynomial Projection Filter Sine}
\usage{
standardInverseSineDer(
  xt,
  N,
  k,
  deltat = 1,
  passInSineMat = NULL,
  returnSineMat = FALSE,
  passInSineUnder = NULL,
  penalty = 1
)
}
\arguments{
\item{xt}{time series}

\item{N}{Total number of observations}

\item{k}{Number of tapers}

\item{deltat}{default is 1 unless using different nyquist}

\item{passInSineMat}{If sine matrix was calculated already}

\item{returnSineMat}{will return the sine taper matrix back to the parent function}

\item{passInSineUnder}{If sine undersampling matrix was calculated already}

\item{penalty}{1 is no penalty , 0.2  would give seq(from =  1, to =  1/(0.2*k, length.out = k)
penalty to each respective taper}
}
\value{
list of $StdInverse and $Freq which are the columns of StdInverse,
if returnSineMat = TRUE then will also return $SineMat
}
\description{
NOTE: this could be sped up by only calculating the fft once!! for all f
}
